Allows to add own paint calls that depend on the transformation matrix by overriding the ScrollbarPainter implementation.
 To do so, override the "void setScrollbarControllers()" method in the
BetterInteractiveViewerBaseState.
This is made possible by encapsulating the "RawTransformScrollbarController getPlatformScrollbarController" method in
auto_platform_scrollbar_controller.dart in a new class "AutoPlatformScrollbarController" that provides all the
necessary methods needed to configure and paint the scrollbars and delegating to the platform specific scrollbar
implementation. To add own paint calls, override the "void paint" method in the AutoPlatformScrollbarController class.